% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-images.R
\name{bioc-images}
\alias{bioc-images}
\alias{BiocFavicon}
\alias{BiocSticker}
\title{Get various Bioconductor images}
\usage{
BiocFavicon()

BiocSticker(mode = c("static", "animated"))
}
\arguments{
\item{mode}{String specifying the type of sticker to show.}
}
\value{
\code{BiocFavicon} will return a path to a \code{favicon.ico} file.

\code{BiocSticker} will return a URL or path to a sticker.
}
\description{
Helper functions to pull down images to use in the book.
These aim to provide a sensible default for Bioconductor-related books.
}
\examples{
BiocFavicon()

BiocSticker()
}
\author{
Aaron Lun
}
