% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseStart.R
\name{collapseStart}
\alias{collapseStart}
\alias{collapseEnd}
\title{Print the collapse opening and ending}
\usage{
collapseStart(message)

collapseEnd()
}
\arguments{
\item{message}{String containing a message to insert in the collapsible header.}
}
\value{
Both functions will \code{\link{cat}} HTML tags; one to start and another to end each collapsible chunk.
}
\description{
Print HTML tags to open and close the collapsible chunks.
}
\examples{
collapseStart("This is collapsible")
cat("something inside the chunk\n")
collapseEnd()

}
\author{
Aaron Lun
}
