% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileBook.R
\name{compileBook}
\alias{compileBook}
\title{Compile the book}
\usage{
compileBook(src.dir, work.dir, final.dir, input = "input.Rmd", ...)
}
\arguments{
\item{input}{String containing the name of the file to pass to \code{\link[bookdown]{render_book}}.}

\item{...}{Further arguments to pass to \code{\link[bookdown]{render_book}}.}

\item{src}{String containing the path to the book Rmarkdown sources.}

\item{work}{String containing the path to the workspace used to compile the book.}

\item{final}{String containing the path to the final location for the compiled book's HTMLs.}
}
\value{
\code{work} is populated with the book sources and intermediate content (e.g., caches).
\code{final} is populated with the HTMLs.
A \code{NULL} is invisibly returned.
}
\description{
Compile a \pkg{bookdown} book in a separate workspace and copy the compiled book to a different location.
}
\seealso{
\code{\link{configureBook}}, where this function is called in the Makefile.

\code{\link{getBookCache}}, typically used to generate a good choice for \code{work}.
}
\author{
Aaron Lun
}
