% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileChapter.R
\name{compileChapter}
\alias{compileChapter}
\title{Compile a Rmarkdown report}
\usage{
compileChapter(path)
}
\arguments{
\item{path}{String containing a path to an Rmarkdown file.}
}
\value{
All files are (re)compiled to generate the corresponding \code{*_cache} directories.
\code{NULL} is invisibly returned.
}
\description{
Compile a report, typically so that \code{\link{extractCached}} calls work correctly in other reports.
}
\details{
Compilation is performed in an isolated session using \code{\link{r}} from the \pkg{callr} package.
This ensures that settings from one chapter do not affect the next chapter.

If an error is encountered during compilation of any Rmarkdown file,
the standard output of \code{\link{render}} leading up to the error is printed out before the function exists.
}
\seealso{
\code{\link{extractCached}}, which calls this function.
}
\author{
Aaron Lun
}
