% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefile.R
\name{createMakefile}
\alias{createMakefile}
\title{Create a compilation Makefile}
\usage{
createMakefile(dir, pattern = "\\\\.Rmd$", ..., fname = "Makefile")
}
\arguments{
\item{fname}{String containing the name of the output Makefile.}
}
\value{
A Makefile is created in \code{dir} with the name \code{fname}
and a \code{NULL} is invisibly returned.
}
\description{
Create a Makefile for compiling individual chapters, potentially in parallel.
This respects the dependencies between chapter as detected by \code{\link{buildReportGraph}}.
}
\details{
This uses the markdown output file as an indicator of successful \code{\link{knit}}ting of a chapter.
}
\author{
Aaron Lun
}
