% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCached.R
\name{extractCached}
\alias{extractCached}
\title{Extract cached objects}
\usage{
extractCached(
  donor,
  chunk,
  objects,
  locations = ".",
  envir = topenv(parent.frame())
)
}
\arguments{
\item{donor}{String containing a regular expression to use to identify the donor Rmarkdown file.}

\item{chunk}{String containing the name of the requested chunk.}

\item{objects}{Character vector containing variable names for one or more objects to be extracted.}

\item{locations}{Character vector containing the possible locations to search for the donor file.}

\item{envir}{Environment where the loaded objects should be stored.}
}
\value{
Variables with names \code{objects} are created in \code{envir}.
A markdown chunk (wrapped in a collapsible element) is printed that contains all commands needed to generate those objects, 
based on the code in the named chunks of the donor Rmarkdown file.
}
\description{
Extract specific R objects from the knitr cache of a previously compiled Rmarkdown file (the \dQuote{donor})
so that it can be used in the compilation process of another Rmarkdown file (the \dQuote{acceptor}).
}
\details{
Each R object is extracted in its state at the requested \code{chunk} and inserted into \code{envir}.
Note that the object does not have to be generated or even referenced in \code{chunk},
provided it was generated in a previous chunk.

The parser in this function is rather limited,
so the donor Rmarkdown file is subject to several constraints:
\itemize{
\item All chunks involved in generating the requested objects (indirectly or otherwise) should be named.
\item All named chunks should be executed, i.e., no \code{eval=FALSE}.
\item All relevant code occurs within triple backticks, i.e., any inline code should be read-only.
\item All triple backticks occur at the start of the line, i.e., no code nested in list elements.
\item Any assignment or modifications to variables are done \emph{correctly} with \code{<-}.
}

Unnamed chunks are allowed but cannot be referenced and will not be shown in the output of this function.
This should not be used for code that might affect variables in the named chunks,
i.e., code in unnamed chunks should be \dQuote{read-only} with respect to variables in the named chunks.
Chunks with names starting with \code{unref-} are considered to be the same as unnamed chunks and will be ignored;
this is useful for figure-generating chunks that need to be referenced inside the donor report.

The donor report is found by searching successive \code{locations} for a file that matches \code{donor}.
It will use the first file that is found in this manner.
A flexible search enables this system to continue operating when the path changes due to \code{\link{setupBookChapters}}.

Obviously, this entire process assumes that donor report has already been compiled with \code{cache=TRUE}.
If not, \code{extractCached} will compile it (and thus generate the cache) using \code{\link{compileCached}}.
}
\seealso{
\code{\link{setupHTML}} and \code{\link{chapterPreamble}}, to set up the code for the collapsible element.

\code{\link{compileChapter}}, to compile a Rmarkdown report to generate the cache.
}
\author{
Aaron Lun
}
