% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBookCache.R
\name{getBookCache}
\alias{getBookCache}
\title{Get the local book cache}
\usage{
getBookCache(package, clear = TRUE)
}
\arguments{
\item{package}{String containing the name of the book package.}

\item{clear}{Logical scalar indicating whether old caches should be wiped.}
}
\value{
String containing the path to the cache directory for this book package.
}
\description{
Get the path to the cache directory in which the book will be built.
}
\details{
The output path contains the version of the specified \code{package}.
If \code{clear=TRUE}, any caches corresponding to older versions of the package are destroyed.

If the environment variable \code{REBOOK_<PKG>_CACHE} is set for some package name \code{PKG} (uppercased with dots replaced by underscores),
this is expected to contain the desired path and is returned directly. 
This functionality should only be used by experienced developers.
}
\examples{
\dontrun{
getBookCache('OSCA.workflows')
}

}
\seealso{
\code{\link{configureBook}}, where this function is used in the Makefile.

\code{\link{extractFromPackage}}, which populates the cache directory if this is not supplied.
}
\author{
Aaron Lun
}
