% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupChapters.R
\name{setupChapters}
\alias{setupChapters}
\title{Set up the book materials}
\usage{
setupChapters(chapters, dir)
}
\arguments{
\item{chapters}{A character vector of paths to Rmarkdown files to be converted to chapters.
Alternatively, a list of such vectors.}

\item{dir}{String containing the location of the output files.}
}
\value{
Creates or updates the Rmarkdown files comprising the book's chapters in \code{dir}.
Returns \code{NULL} invisibly.
}
\description{
Convert Rmarkdown files to numbered chapters for \pkg{bookdown} compilation.
}
\details{
If \code{chapters} is a character vector with contents \code{c("first", "second", ...)}
chapters are created in the manner of \code{01-first.Rmd}, \code{02-second.Rmd} and so on.

If \code{chapters} is a list of character vectors,
each vector is assumed to represent a \dQuote{Part} of the book and each file is a chapter in that Part.
In this case, chapters would look like \code{P1_W01.first.Rmd}, \code{P1_W02.second.Rmd} and so on.

All existing files in \code{dir} that follow a similar naming format are destroyed,
to avoid any confusion between the old and new files.
}
\author{
Aaron Lun
}
