% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureBook.R
\name{configureBook}
\alias{configureBook}
\title{Helper configuration function for books}
\usage{
configureBook(prefix = NULL, input = "index.Rmd")
}
\arguments{
\item{prefix}{String containing the prefix to be used when \code{\link{link}}ing from other books.}

\item{input}{See \code{\link{scrapeReferences}}.}
}
\value{
\code{NULL}, invisibly.
A \code{"references.csv"} file is created in the \code{inst/rebook} directory.
If \code{prefix} is specified, a \code{"prefix.csv"} file is also created.
}
\description{
Helper function to run in the \code{configure} of Bioconductor books,
to set up for \code{\link{link}}ing from other books.
}
\details{
This assumes that the \pkg{bookdown}-formatted book is located at \code{inst/book}.
}
\seealso{
\code{\link{scrapeReferences}}, which is called by this function to create the reference table.

\code{\link{link}}, which is used by other books to link to the \code{configure}d book.
}
\author{
Aaron Lun
}
