% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefile.R
\name{createMakefile}
\alias{createMakefile}
\title{Create a compilation Makefile}
\usage{
createMakefile(dir = ".", pattern = "\\\\.Rmd$", ..., fname = "Makefile")
}
\arguments{
\item{dir}{String containing the path to the directory containing Rmarkdown reports.
This is searched recursively for all files ending in \code{".Rmd"}.}

\item{pattern}{Further arguments to pass to \code{\link{list.files}} 
when searching for Rmarkdown reports.}

\item{fname}{String containing the name of the output Makefile.}
}
\value{
A Makefile is created in \code{dir} with the name \code{fname}
and a \code{NULL} is invisibly returned.
}
\description{
Create a Makefile for compiling individual chapters, potentially in parallel.
This respects the dependencies between chapter as detected by \code{\link{buildChapterGraph}}.
}
\details{
This uses the markdown output file as an indicator of successful \code{\link{knit}}ting of a chapter.
Caches are left in the current working directory after the compilation of each report.
It is assumed that \pkg{bookdown}'s \code{render_book} is smart enough to find and use these caches.
}
\examples{
dir <- tempfile()
dir.create(dir)

tmp1 <- file.path(dir, "alpha.Rmd")
write(file=tmp1, "```{r, echo=FALSE, results='asis'}
rebook::chapterPreamble()
```

```{r}
rodan <- 1
```")

tmp2 <- file.path(dir, "bravo.Rmd")
write(file=tmp2, "```{r, echo=FALSE, results='asis'}
rebook::chapterPreamble()
```

```{r}
extractCached('alpha.Rmd')
```")

# Creating the Makefile:
createMakefile(dir)
cat(readLines(file.path(dir, "Makefile")), sep="\n")

}
\author{
Aaron Lun
}
