% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDependencies.R
\name{updateDependencies}
\alias{updateDependencies}
\title{Update the dependencies}
\usage{
updateDependencies(dir = ".", extra = NULL, ...)
}
\arguments{
\item{dir}{String containing the path to the directory containing the book \code{DESCRIPTION} file.}

\item{extra}{Character vector of extra packages to be added to imports,
usually from packages that are in \code{Suggests} and thus not caught directly by \code{\link{scrapeDependencies}}.}

\item{...}{Further arguments to pass to \code{\link{scrapeDependencies}}.}
}
\value{
The \code{DESCRIPTION} file in \code{dir} is updated.
}
\description{
Update the \code{Imports} field in the book's \code{DESCRIPTION} file with the latest dependencies.
}
\author{
Aaron Lun
}
